-Added intro screen -Made some changes to make code more compatible with 64bit -Added code to db layer for simple migrations -moved some code around in BaseViewController -added code for getting ready for the addition of hero,zone, and monster and all that shit. -added Story constants

+++ HabitRPG2/DailyHelper.h
@@ -15,7 +15,7 @@
 @interface DailyHelper : NSObject
 @property (nonatomic,strong) CommonUtilities *commonHelper;
 -(BOOL)isDailyCompleteForTheDay:(Daily *)daily;
--(NSDate *)calculateNextDueTime:(NSDate *)checkinDate WithRate:(int)rate;
+-(NSDate *)calculateNextDueTime:(NSDate *)checkinDate WithRate:(long)rate;
 -(int)calculateActiveDaysHash:(NSMutableArray *)activeDays;
 -(void)setActiveDaySwitches:(NSMutableArray *)activeDays fromHash:(NSInteger)hash;
 @end
